+++ discoverspec/discoverapi.pb.go
@@ -29,91 +29,6 @@ const (
 // of the legacy proto package is being used.
 const _ = proto.ProtoPackageIsVersion4
 
-type ListApplicationIdentitiesRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-}
-
-func (x *ListApplicationIdentitiesRequest) Reset() {
-	*x = ListApplicationIdentitiesRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *ListApplicationIdentitiesRequest) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*ListApplicationIdentitiesRequest) ProtoMessage() {}
-
-func (x *ListApplicationIdentitiesRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use ListApplicationIdentitiesRequest.ProtoReflect.Descriptor instead.
-func (*ListApplicationIdentitiesRequest) Descriptor() ([]byte, []int) {
-	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{0}
-}
-
-type ListApplicationIdentitiesResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	Identities []*Identity `protobuf:"bytes,1,rep,name=identities,proto3" json:"identities,omitempty"`
-}
-
-func (x *ListApplicationIdentitiesResponse) Reset() {
-	*x = ListApplicationIdentitiesResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
-}
-
-func (x *ListApplicationIdentitiesResponse) String() string {
-	return protoimpl.X.MessageStringOf(x)
-}
-
-func (*ListApplicationIdentitiesResponse) ProtoMessage() {}
-
-func (x *ListApplicationIdentitiesResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		if ms.LoadMessageInfo() == nil {
-			ms.StoreMessageInfo(mi)
-		}
-		return ms
-	}
-	return mi.MessageOf(x)
-}
-
-// Deprecated: Use ListApplicationIdentitiesResponse.ProtoReflect.Descriptor instead.
-func (*ListApplicationIdentitiesResponse) Descriptor() ([]byte, []int) {
-	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{1}
-}
-
-func (x *ListApplicationIdentitiesResponse) GetIdentities() []*Identity {
-	if x != nil {
-		return x.Identities
-	}
-	return nil
-}
-
 type WatchRequest struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -123,7 +38,7 @@ type WatchRequest struct {
 func (x *WatchRequest) Reset() {
 	*x = WatchRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2]
+		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[0]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -136,7 +51,7 @@ func (x *WatchRequest) String() string {
 func (*WatchRequest) ProtoMessage() {}
 
 func (x *WatchRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2]
+	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[0]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -149,7 +64,7 @@ func (x *WatchRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use WatchRequest.ProtoReflect.Descriptor instead.
 func (*WatchRequest) Descriptor() ([]byte, []int) {
-	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{2}
+	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{0}
 }
 
 type WatchResponse struct {
@@ -157,13 +72,17 @@ type WatchResponse struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Available bool `protobuf:"varint,1,opt,name=available,proto3" json:"available,omitempty"`
+	// Application is the identity of the Dogma application that the
+	Application *Identity `protobuf:"bytes,1,opt,name=application,proto3" json:"application,omitempty"`
+	// Available indicates whether the application should currently be considered
+	// available or not.
+	Available bool `protobuf:"varint,2,opt,name=available,proto3" json:"available,omitempty"`
 }
 
 func (x *WatchResponse) Reset() {
 	*x = WatchResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[3]
+		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -176,7 +95,7 @@ func (x *WatchResponse) String() string {
 func (*WatchResponse) ProtoMessage() {}
 
 func (x *WatchResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[3]
+	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -189,7 +108,14 @@ func (x *WatchResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use WatchResponse.ProtoReflect.Descriptor instead.
 func (*WatchResponse) Descriptor() ([]byte, []int) {
-	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{3}
+	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{1}
+}
+
+func (x *WatchResponse) GetApplication() *Identity {
+	if x != nil {
+		return x.Application
+	}
+	return nil
 }
 
 func (x *WatchResponse) GetAvailable() bool {
@@ -199,22 +125,22 @@ func (x *WatchResponse) GetAvailable() bool {
 	return false
 }
 
-// Identity represents the identity of an application or handler.
+// Identity represents the identity of an application.
 type Identity struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	// Name is the entity's unique name.
+	// Name is the application's unique name.
 	Name string `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
-	// Key is the entity's immutable, unique key.
+	// Key is the application's immutable, unique key.
 	Key string `protobuf:"bytes,2,opt,name=key,proto3" json:"key,omitempty"`
 }
 
 func (x *Identity) Reset() {
 	*x = Identity{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[4]
+		mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -227,7 +153,7 @@ func (x *Identity) String() string {
 func (*Identity) ProtoMessage() {}
 
 func (x *Identity) ProtoReflect() protoreflect.Message {
-	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[4]
+	mi := &file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -240,7 +166,7 @@ func (x *Identity) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use Identity.ProtoReflect.Descriptor instead.
 func (*Identity) Descriptor() ([]byte, []int) {
-	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{4}
+	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescGZIP(), []int{2}
 }
 
 func (x *Identity) GetName() string {
@@ -265,43 +191,29 @@ var file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDesc 
 	0x63, 0x2f, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x73, 0x70, 0x65, 0x63, 0x2f, 0x64,
 	0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x61, 0x70, 0x69, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
 	0x12, 0x19, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e,
-	0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x22, 0x22, 0x0a, 0x20, 0x4c,
-	0x69, 0x73, 0x74, 0x41, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x64,
-	0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22,
-	0x68, 0x0a, 0x21, 0x4c, 0x69, 0x73, 0x74, 0x41, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74, 0x69,
-	0x6f, 0x6e, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x43, 0x0a, 0x0a, 0x69, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x69,
-	0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32, 0x23, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61,
-	0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63,
-	0x6f, 0x76, 0x65, 0x72, 0x2e, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x79, 0x52, 0x0a, 0x69,
-	0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x22, 0x0e, 0x0a, 0x0c, 0x57, 0x61, 0x74,
-	0x63, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x2d, 0x0a, 0x0d, 0x57, 0x61, 0x74,
-	0x63, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x1c, 0x0a, 0x09, 0x61, 0x76,
-	0x61, 0x69, 0x6c, 0x61, 0x62, 0x6c, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08, 0x52, 0x09, 0x61,
-	0x76, 0x61, 0x69, 0x6c, 0x61, 0x62, 0x6c, 0x65, 0x22, 0x30, 0x0a, 0x08, 0x49, 0x64, 0x65, 0x6e,
-	0x74, 0x69, 0x74, 0x79, 0x12, 0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x32, 0x84, 0x02, 0x0a, 0x0b, 0x44,
-	0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x41, 0x50, 0x49, 0x12, 0x96, 0x01, 0x0a, 0x19, 0x4c,
-	0x69, 0x73, 0x74, 0x41, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x49, 0x64,
-	0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x12, 0x3b, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61,
-	0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63,
-	0x6f, 0x76, 0x65, 0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x41, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61,
-	0x74, 0x69, 0x6f, 0x6e, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x3c, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65,
-	0x72, 0x2e, 0x4c, 0x69, 0x73, 0x74, 0x41, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f,
-	0x6e, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x69, 0x65, 0x73, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x5c, 0x0a, 0x05, 0x57, 0x61, 0x74, 0x63, 0x68, 0x12, 0x27, 0x2e, 0x64,
-	0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e,
-	0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x2e, 0x57, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x28, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e,
-	0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65,
-	0x72, 0x2e, 0x57, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30,
-	0x01, 0x42, 0x2e, 0x5a, 0x2c, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f,
-	0x64, 0x6f, 0x67, 0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70,
-	0x73, 0x70, 0x65, 0x63, 0x2f, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x73, 0x70, 0x65,
-	0x63, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x22, 0x0e, 0x0a, 0x0c, 0x57,
+	0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x74, 0x0a, 0x0d, 0x57,
+	0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x45, 0x0a, 0x0b,
+	0x61, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x01, 0x20, 0x01, 0x28,
+	0x0b, 0x32, 0x23, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f,
+	0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x2e, 0x49, 0x64,
+	0x65, 0x6e, 0x74, 0x69, 0x74, 0x79, 0x52, 0x0b, 0x61, 0x70, 0x70, 0x6c, 0x69, 0x63, 0x61, 0x74,
+	0x69, 0x6f, 0x6e, 0x12, 0x1c, 0x0a, 0x09, 0x61, 0x76, 0x61, 0x69, 0x6c, 0x61, 0x62, 0x6c, 0x65,
+	0x18, 0x02, 0x20, 0x01, 0x28, 0x08, 0x52, 0x09, 0x61, 0x76, 0x61, 0x69, 0x6c, 0x61, 0x62, 0x6c,
+	0x65, 0x22, 0x30, 0x0a, 0x08, 0x49, 0x64, 0x65, 0x6e, 0x74, 0x69, 0x74, 0x79, 0x12, 0x12, 0x0a,
+	0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e, 0x61, 0x6d,
+	0x65, 0x12, 0x10, 0x0a, 0x03, 0x6b, 0x65, 0x79, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03,
+	0x6b, 0x65, 0x79, 0x32, 0x6b, 0x0a, 0x0b, 0x44, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x41,
+	0x50, 0x49, 0x12, 0x5c, 0x0a, 0x05, 0x57, 0x61, 0x74, 0x63, 0x68, 0x12, 0x27, 0x2e, 0x64, 0x6f,
+	0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64,
+	0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x2e, 0x57, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x71,
+	0x75, 0x65, 0x73, 0x74, 0x1a, 0x28, 0x2e, 0x64, 0x6f, 0x67, 0x6d, 0x61, 0x2e, 0x69, 0x6e, 0x74,
+	0x65, 0x72, 0x6f, 0x70, 0x2e, 0x76, 0x31, 0x2e, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72,
+	0x2e, 0x57, 0x61, 0x74, 0x63, 0x68, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x30, 0x01,
+	0x42, 0x2e, 0x5a, 0x2c, 0x67, 0x69, 0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x64,
+	0x6f, 0x67, 0x6d, 0x61, 0x74, 0x69, 0x71, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6f, 0x70, 0x73,
+	0x70, 0x65, 0x63, 0x2f, 0x64, 0x69, 0x73, 0x63, 0x6f, 0x76, 0x65, 0x72, 0x73, 0x70, 0x65, 0x63,
+	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -316,22 +228,18 @@ func file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDesc
 	return file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDescData
 }
 
-var file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
+var file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
 var file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_goTypes = []interface{}{
-	(*ListApplicationIdentitiesRequest)(nil),  // 0: dogma.interop.v1.discover.ListApplicationIdentitiesRequest
-	(*ListApplicationIdentitiesResponse)(nil), // 1: dogma.interop.v1.discover.ListApplicationIdentitiesResponse
-	(*WatchRequest)(nil),                      // 2: dogma.interop.v1.discover.WatchRequest
-	(*WatchResponse)(nil),                     // 3: dogma.interop.v1.discover.WatchResponse
-	(*Identity)(nil),                          // 4: dogma.interop.v1.discover.Identity
+	(*WatchRequest)(nil),  // 0: dogma.interop.v1.discover.WatchRequest
+	(*WatchResponse)(nil), // 1: dogma.interop.v1.discover.WatchResponse
+	(*Identity)(nil),      // 2: dogma.interop.v1.discover.Identity
 }
 var file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_depIdxs = []int32{
-	4, // 0: dogma.interop.v1.discover.ListApplicationIdentitiesResponse.identities:type_name -> dogma.interop.v1.discover.Identity
-	0, // 1: dogma.interop.v1.discover.DiscoverAPI.ListApplicationIdentities:input_type -> dogma.interop.v1.discover.ListApplicationIdentitiesRequest
-	2, // 2: dogma.interop.v1.discover.DiscoverAPI.Watch:input_type -> dogma.interop.v1.discover.WatchRequest
-	1, // 3: dogma.interop.v1.discover.DiscoverAPI.ListApplicationIdentities:output_type -> dogma.interop.v1.discover.ListApplicationIdentitiesResponse
-	3, // 4: dogma.interop.v1.discover.DiscoverAPI.Watch:output_type -> dogma.interop.v1.discover.WatchResponse
-	3, // [3:5] is the sub-list for method output_type
-	1, // [1:3] is the sub-list for method input_type
+	2, // 0: dogma.interop.v1.discover.WatchResponse.application:type_name -> dogma.interop.v1.discover.Identity
+	0, // 1: dogma.interop.v1.discover.DiscoverAPI.Watch:input_type -> dogma.interop.v1.discover.WatchRequest
+	1, // 2: dogma.interop.v1.discover.DiscoverAPI.Watch:output_type -> dogma.interop.v1.discover.WatchResponse
+	2, // [2:3] is the sub-list for method output_type
+	1, // [1:2] is the sub-list for method input_type
 	1, // [1:1] is the sub-list for extension type_name
 	1, // [1:1] is the sub-list for extension extendee
 	0, // [0:1] is the sub-list for field type_name
@@ -344,30 +252,6 @@ func file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_init() 
 	}
 	if !protoimpl.UnsafeEnabled {
 		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListApplicationIdentitiesRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ListApplicationIdentitiesResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*WatchRequest); i {
 			case 0:
 				return &v.state
@@ -379,7 +263,7 @@ func file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_init() 
 				return nil
 			}
 		}
-		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*WatchResponse); i {
 			case 0:
 				return &v.state
@@ -391,7 +275,7 @@ func file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_init() 
 				return nil
 			}
 		}
-		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*Identity); i {
 			case 0:
 				return &v.state
@@ -410,7 +294,7 @@ func file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_init() 
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_github_com_dogmatiq_interopspec_discoverspec_discoverapi_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   5,
+			NumMessages:   3,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
@@ -436,12 +320,8 @@ const _ = grpc.SupportPackageIsVersion6
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
 type DiscoverAPIClient interface {
-	// ListApplicationIdentities returns the identity of all applications.
-	ListApplicationIdentities(ctx context.Context, in *ListApplicationIdentitiesRequest, opts ...grpc.CallOption) (*ListApplicationIdentitiesResponse, error)
-	// Watch starts watching the server for updates to its availability.
-	//
-	// The client SHOULD treat the server as though it had not been discovered at
-	// all until it sends a WatchResponse with the "available" field set to true.
+	// Watch starts watching the server for updates to the availability of Dogma
+	// applications.
 	Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (DiscoverAPI_WatchClient, error)
 }
 
@@ -451,15 +331,6 @@ type discoverAPIClient struct {
 
 func NewDiscoverAPIClient(cc grpc.ClientConnInterface) DiscoverAPIClient {
 	return &discoverAPIClient{cc}
-}
-
-func (c *discoverAPIClient) ListApplicationIdentities(ctx context.Context, in *ListApplicationIdentitiesRequest, opts ...grpc.CallOption) (*ListApplicationIdentitiesResponse, error) {
-	out := new(ListApplicationIdentitiesResponse)
-	err := c.cc.Invoke(ctx, "/dogma.interop.v1.discover.DiscoverAPI/ListApplicationIdentities", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
 }
 
 func (c *discoverAPIClient) Watch(ctx context.Context, in *WatchRequest, opts ...grpc.CallOption) (DiscoverAPI_WatchClient, error) {
@@ -496,12 +367,8 @@ func (x *discoverAPIWatchClient) Recv() (*WatchResponse, error) {
 
 // DiscoverAPIServer is the server API for DiscoverAPI service.
 type DiscoverAPIServer interface {
-	// ListApplicationIdentities returns the identity of all applications.
-	ListApplicationIdentities(context.Context, *ListApplicationIdentitiesRequest) (*ListApplicationIdentitiesResponse, error)
-	// Watch starts watching the server for updates to its availability.
-	//
-	// The client SHOULD treat the server as though it had not been discovered at
-	// all until it sends a WatchResponse with the "available" field set to true.
+	// Watch starts watching the server for updates to the availability of Dogma
+	// applications.
 	Watch(*WatchRequest, DiscoverAPI_WatchServer) error
 }
 
@@ -509,33 +376,12 @@ type DiscoverAPIServer interface {
 type UnimplementedDiscoverAPIServer struct {
 }
 
-func (*UnimplementedDiscoverAPIServer) ListApplicationIdentities(context.Context, *ListApplicationIdentitiesRequest) (*ListApplicationIdentitiesResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method ListApplicationIdentities not implemented")
-}
 func (*UnimplementedDiscoverAPIServer) Watch(*WatchRequest, DiscoverAPI_WatchServer) error {
 	return status.Errorf(codes.Unimplemented, "method Watch not implemented")
 }
 
 func RegisterDiscoverAPIServer(s *grpc.Server, srv DiscoverAPIServer) {
 	s.RegisterService(&_DiscoverAPI_serviceDesc, srv)
-}
-
-func _DiscoverAPI_ListApplicationIdentities_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(ListApplicationIdentitiesRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(DiscoverAPIServer).ListApplicationIdentities(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/dogma.interop.v1.discover.DiscoverAPI/ListApplicationIdentities",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(DiscoverAPIServer).ListApplicationIdentities(ctx, req.(*ListApplicationIdentitiesRequest))
-	}
-	return interceptor(ctx, in, info, handler)
 }
 
 func _DiscoverAPI_Watch_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -562,12 +408,7 @@ func (x *discoverAPIWatchServer) Send(m *WatchResponse) error {
 var _DiscoverAPI_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "dogma.interop.v1.discover.DiscoverAPI",
 	HandlerType: (*DiscoverAPIServer)(nil),
-	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "ListApplicationIdentities",
-			Handler:    _DiscoverAPI_ListApplicationIdentities_Handler,
-		},
-	},
+	Methods:     []grpc.MethodDesc{},
 	Streams: []grpc.StreamDesc{
 		{
 			StreamName:    "Watch",