+++ CHANGELOG.md
@@ -9,6 +9,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.4.0] - 2020-12-22
+
+### Add
+
+- **[BC]** Add `discoverspec.DiscoverAPI.WatchResponse.Application` field
+
+### Removed
+
+- **[BC]** Remove `discoverspec.DiscoverAPI.ListApplicationIdentities()`
+
 ## [0.3.0] - 2020-12-22
 
 ### Added
@@ -41,6 +51,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.2.0]: https://github.com/dogmatiq/testkit/releases/tag/v0.2.0
 [0.2.1]: https://github.com/dogmatiq/testkit/releases/tag/v0.2.1
 [0.3.0]: https://github.com/dogmatiq/testkit/releases/tag/v0.3.0
+[0.4.0]: https://github.com/dogmatiq/testkit/releases/tag/v0.4.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD